refactor(modems): split modem-update-loop into modules
